refactor: move httputils directly under pkg
- Remove Push method from responseWriterWithCode
- Remove secrets volume in the deploy action
