+++ internal/rest/middleware.go
@@ -9,7 +9,7 @@ import (
 	"time"
 
 	"github.com/shivanshkc/squelette/internal/logger"
-	"github.com/shivanshkc/squelette/pkg/utils/httputils"
+	"github.com/shivanshkc/squelette/pkg/httputils"
 
 	"github.com/google/uuid"
 )
